+++ TIGLCreator/src/modificators/ModificatorWingsWidget.h
@@ -34,7 +34,8 @@ class ModificatorWingsWidget : public ModificatorWidget
     Q_OBJECT
 
 signals:
-    void undoCommandRequired();
+    void addWingRequested();
+    void deleteWingRequested();
 
 public slots:
 
@@ -45,12 +46,8 @@ public:
     explicit ModificatorWingsWidget(QWidget *parent = nullptr);
     ~ModificatorWingsWidget();
 
-    void setWings(tigl::CCPACSWings& wings, ProfilesDBManager* profilesDB);
-
 private:
     Ui::ModificatorWingsWidget *ui;
-    tigl::CCPACSWings* wings;
-    ProfilesDBManager* profilesDB;
 };
 
 #endif // MODIFICATORWINGSWIDGET_H